refactor: move phone helper to global support namespace
